Zero Two (#2)
* Solves D02P1

This.. was a lot faster than D01P1, that's for sure!

* D02P2 solved too

This was much much faster than D01P*. Gathered a couple of points on the
private leaderboard I'm a part of too... so.. bonus!

I had started preparing for a more... complex password policy, but nope, that
wasn't needed at all. Leaving in the commented out code though.