One three (#6)
* Solves D13P1, moves D12 into unit tests.

Initially, I thought I might reuse D01's ZOTree structure, since the puzzle
could have been expressed in a variation of the subtraction method, although
that might have been slightly annoying to implement. Just stuck with some basic
processing, although learning how to use reduce with dictionaries was cool.

* Stops ignoring unconstrained buses

Part 2 needs to know how many unconstrained buses exist between the constrained
ones, so that the sequential departure times can be caluclated.

* Tweaks data 'structure' used

This is anticipation of a P2 rework, namely finding the largest departure
interval, and using that for the iteration of sequential/staggered departure
times, instead of using the first bus' interval.

* Tests an iteration of the largest departure time interval loop

It's still in the infancy, so the wrapping loop to continuously search is not
pulled into the Schedule class. Soon (tm).

* Halts my primitive D13P2 implementation

It was running in the background, like the music, while I stared at a piece of
paper, lost in trying to figure out a way out of this endless loop (see the
`interval.sh` script at the root.

I has a dumb feel now, and I .. don't like it. Why couldn't I have thought of
something like that animation? See the script for the link to the anim.

* Solves D13P2 with animation assistance.

Seriously, that.. this is the solution? A measly 10-ish lines of code, at least
in my implementation - the working portion of the solution anyways.

And the starting at timestamp suggestion? Pointless, at least, with this kind
of algorithm.