Adds the files required for execution by others (#8)
* Adds the previously ignored files, where appropriate

Basically, the input data, in both the `example.txt` and `generated.txt` files,
as well as the results, supported by the Python module and JSON data files.

These are used in support of the active puzzle being solved, as well as using
similar calls for the unit tests. Turns out, there wasn't much reuse of common
items, such as data structures, so the UTs really didn't come into play. But it
is still a nice thing to know, if at some point, I come back to continue work
on D22P2 and other puzzles not solved during 2020.

* Adds D22P1 as unit tests, skipping P2 test calls.

The launcher file was updated to only call unit tests, but the code was merely
commented out, because of the reason listed therein: I don't want to look over
Git history to figure out how to do it, although, this is a weak reason, it is
mine to make.

Also forgot to add the result of the generated data, so fixed that (found via
unit tests, oddly enough). Learned how to skip unit tests in Python, so not all
was lost. But this will pretty much wrap up my AoC 2020 repository contribs.

Will likely read-only-lock the repo, and leave it available for .. let's face
it: laughs/cries in the future. Anyways, despite some frustration, it's been
fun. Mostly.